Add constants, middleware, and user management functionality with socket integration
