+++ backend/index.js
@@ -1,10 +1,10 @@
 import express from "express";
-import mongoose from "mongoose";
 import dotenv from "dotenv";
 import cookieParser from "cookie-parser";
 import authRoutes from "./modules/auth/auth.routes.js";
 import connectToDb from "./db.js";
-import { app, server } from "socket.io/socket.main.js";
+import { app, server } from "./socket/socket.main.js";
+
 dotenv.config({ path: "./.env" });
 // const app = express();
 
@@ -16,7 +16,11 @@ await connectToDb();
 
 app.use("/auth", authRoutes);
 
+app.get("/", (_req, res) => {
+  res.send(`Server is ready. Testing Nodemon`);
+});
+
 const PORT = process.env.PORT || 3000;
 server.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
+  console.log(`Server running on port http://localhost:${PORT}`);
 });
