+++ backend/models/conversation.model.js
@@ -6,18 +6,17 @@ const conversationSchema = new Schema(
       {
         type: Schema.Types.ObjectId,
         ref: "User",
-        required: true,
       },
     ],
     messages: [
       {
         type: Schema.Types.ObjectId,
         ref: "Message",
-        required: true,
+        default: [],
       },
     ],
   },
-  { timestamps: true },
+  { timestamps: true }
 );
 
 const Conversation = model("Conversation", conversationSchema);
