+++ backend/modules/auth/auth.controller.js
@@ -1,10 +1,4 @@
-import {
-  loginUser,
-  userData,
-  logoutUser,
-  signupUser,
-  userByEmail,
-} from "./auth.service.js";
+import { loginUser, logoutUser, signupUser } from "./auth.service.js";
 
 export const signup = (req, res) => {
   signupUser(req, res);
@@ -14,14 +8,6 @@ export const login = (req, res) => {
   loginUser(req, res);
 };
 
-export const getUserData = (req, res) => {
-  userData(req, res);
-};
-
 export const logout = (req, res) => {
   logoutUser(req, res);
-};
-
-export const getUserByEmail = (req, res) => {
-  userByEmail(req, res);
 };
