+++ backend/modules/user/user.routes.js
@@ -1,0 +1,5 @@
+import checkToken from "../../middlewares/checkToken";
+import { getUserByEmail, getUserData } from "./user.controller.js";
+
+router.get("/user", checkToken, getUserData);
+router.get("/user/email", checkToken, getUserByEmail);
